External example by user can be deleted his examples

+++ app/helpers/external_examples_helper.rb
@@ -1,7 +1,12 @@
 module ExternalExamplesHelper
-  def delete_external_example(id, kata)
+  NO_LINK = ''
+
+  def delete_external_example(id, kata, user)
     text = t(:delete)
-    link_to(text, delete_external_example_path(id, kata), { method: :delete, class: "btn btn-danger"})
+    return link_to(text, delete_external_example_path(id, kata), { method: :delete, class: "btn btn-danger"}) if admin_signed_in?
+    return link_to(text, delete_external_example_path(id, kata), { method: :delete, class: "btn btn-danger"}) if example_belongs?(user)
+
+    NO_LINK
   end
 
   def edit_external_example(id)
